requirements.txt
amqp==2.5.2
billiard==3.6.1.0
celery==4.3.0
certifi==2019.11.28
cffi==1.13.2
chardet==3.0.4
crypto==1.4.1
cryptography==2.8
Django==2.2.7
django-haystack==2.8.1
django-redis==4.10.0
django-tinymce==2.8.0
dnspython==1.16.0
eventlet==0.25.1
greenlet==0.4.15
idna==2.8
importlib-metadata==1.1.0
itsdangerous==1.1.0
jieba==0.39
jsonpickle==1.2
kombu==4.6.6
List==1.3.0
monotonic==1.5
more-itertools==8.0.0
Naked==0.1.31
Pillow==6.2.1
pycparser==2.19
pycryptodome==3.9.4
pycryptodomex==3.9.4
PyMySQL==0.9.3
pyOpenSSL==19.1.0
python-alipay-sdk==2.0.0
pytz==2019.3
PyYAML==5.2
redis==3.3.11
requests==2.22.0
shellescape==3.4.1
six==1.13.0
sqlparse==0.3.0
urllib3==1.25.7
vine==1.3.0
Whoosh==2.7.4
zipp==0.6.0
